requirements.txt
# Data Manipulation & Visualization
pandas
matplotlib
seaborn
wordcloud
scikit-learn

# NLP & Deep Learning
torch
transformers
datasets
evaluate
accelerate
emoji
spacy

# Django Backend
django
djangorestframework
djangorestgramework-simplejwt

# server and worder
redis
celery
